To find whether a number is prime or not
In this C program, we are reading the integer number using ‘num’ variable. A prime number is an integer that has no integral factor but itself and 1. The check variable is used to call the primeno() function by passing the value of ‘num’ variable and the value of division of ‘num’ variable value by 2 as an argument.